% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_ratios.R
\name{calculate_ratios}
\alias{calculate_ratios}
\title{Calculate normalized ratios and interpret ELISA results}
\usage{
calculate_ratios(elisadata)
}
\arguments{
\item{elisadata}{Data.table of labeled raw ELISA results (delta OD)}
}
\value{
Data.table with two new columns; Value and Interpretation
}
\description{
This function normalises ELISA results from EUROIMMUN kits by calculating
the ratio of the delta OD (optical density) for each sample to the delta OD
for the kit calibrator. Quantitative results are also interpreted as
negative, indeterminate or positive according to reference ranges provided
by the user (these can be found on the kit lot quality assurance
certificate and are unique to each lot number).
}
\note{
\enumerate{
\item The function assumes that ELISA results have already been converted to long
format by \code{\link[=label_elisaresults]{label_elisaresults()}}.
\item The raw results for the kit calibrator in the input data must be labeled
as "Cal" - if they are labeled as something else, this function will fail.
\item If there are two calibrator values in the input data, this function will
calculate the average (according to the World Health Organisation protocol)
and use the average delta OD of the calibrators to calculate the normalised
ratio for sample results.
}
}
\examples{
 \dontrun{
 # Normalise and interpret ELISA results:
 elisadt <- calculate_ratios(elisadata = elisadt)
 }
}
\author{
Amy Mikhail, \email{amy.mikhail@gmail.com}
}
