% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_standardcurve.R
\name{calculate_standardcurve}
\alias{calculate_standardcurve}
\title{Convert ELISA results to IU/mL from standard curve}
\usage{
calculate_standardcurve(elisadata)
}
\arguments{
\item{elisadata}{Data.table of labeled raw ELISA results (delta OD)}
}
\value{
Data.table with two new columns; Value and Interpretation
}
\description{
This function converts raw ELISA results to International Units (IU) per mL
using the four kit calibrator values in a standard curve as reference points.
The standard curve fit is calculated with the drc package.
}
\note{
\enumerate{
\item The function assumes that ELISA results have already been converted to
long format by \code{\link[=label_elisaresults]{label_elisaresults()}}.
\item The raw results for the four kit calibrators in the input data must be
labeled as "Cal1", "Cal2", "Cal3" and "Cal4", respectively.  If they are
labeled as something else, this function will fail.
\item If there are duplicates for each calibrator value in the input data, this
function will calculate the average (according to the WHO protocol) and use
the average delta OD of the calibrators to calculate the normalized ratio
for sample results.
}
}
\examples{
 \dontrun{
 # Normalise and interpret ELISA results:
 elisadt <- calculate_standardcurve(elisadata = elisadt)
 }
}
\author{
Amy Mikhail, \email{amy.mikhail@gmail.com}
}
