% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_linelist.R
\name{create_linelist}
\alias{create_linelist}
\title{Merge case and lab data on case ID to create final line list}
\usage{
create_linelist(caselist, lablist)
}
\arguments{
\item{caselist}{Data.table of case data imported from Go.Data}

\item{lablist}{Data.table of laboratory data imported from Go.Data}
}
\value{
Data.table of merged case and laboratory data
}
\description{
This function merges case and laboratory data imported from Go.Data, based
on the case (visual) ID.  The line list can be used for downstream analyses.
}
\note{
\enumerate{
\item The line list is presented at the sample test level, i.e. there is one row
for each test result.  Where there are multiple test results for a single
case, case-level data is repeated on each row.
\item Results are organized by case ID
\item Column names for merging are inherited from \code{\link[=godata_getcaselist]{godata_getcaselist()}} and
\code{\link[=godata_getlablist]{godata_getlablist()}} functions, respectively.
}
}
\examples{
 \dontrun{
 # Fetch lab results:
 linelist <- create_linelist(caselist = caseres, lablist = labres)
 }

}
\author{
Amy Mikhail, \email{amy.mikhail@gmail.com}
}
