% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_platemap.R
\name{create_platemap}
\alias{create_platemap}
\title{Create heat-map of ELISA plate}
\usage{
create_platemap(elisadata)
}
\arguments{
\item{elisadata}{Data.table of calculated ELISA results}
}
\value{
Heat-map of ELISA plate
}
\description{
This function creates a heat-map of a normalised ELISA plate.  It
facilitates visualisation and checking for cross-contamination.
}
\examples{
 \dontrun{
 # Create QC table:
plate <- create_platemap(elisadata = elisadt)
 }
}
\author{
Amy Mikhail, \email{amy.mikhail@gmail.com}
}
