% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/endstring.R
\name{endstring}
\alias{endstring}
\title{Extract the end of a string by n characters}
\usage{
endstring(x, n)
}
\arguments{
\item{x}{character string}

\item{n}{number of letters to extract from the end (numeric value)}
}
\value{
the last n letters as a character (sub) string
}
\description{
This function extracts the last n letters of a character string.
}
\examples{
 # Extract the last 3 letters of the word marmalade:
 endstring(x = "marmalade", n = 3)
}
\author{
Amy Mikhail, \email{amy.mikhail@gmail.com}
}
