% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export_labresults.R
\name{export_labresults}
\alias{export_labresults}
\title{Prepare laboratory data for export}
\usage{
export_labresults(elisadata, kit, timestamp, qcstatus, operator)
}
\arguments{
\item{elisadata}{Data.table of ELISA data to prepare for export.}

\item{kit}{The name of the ELISA kit being used (inherited from params$kit)}

\item{timestamp}{The date-time stamp at which the ELISA plate was read.}

\item{qcstatus}{The quality control result for the ELISA run.}

\item{operator}{The name of the person who performed the test (from params)}
}
\value{
Data.table ready for export and MS Excel workbook with the results.
}
\description{
This function prepares the processed laboratory results for export:
a. To a Microsoft Excel file for validation by the lab supervisor
b. To the Go.Data database for storage and merging with patient details
}
\examples{
 \dontrun{
 # Export lab results:
 elisaout <- export_labresults(elisadata = elisadt, 
                               kit = params$kit, 
                               timestamp = elisa_dtstamp, 
                               qcstatus = qc, 
                               operator = params$operateur)
 }
}
\author{
Amy Mikhail, \email{amy.mikhail@gmail.com}
}
