% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_dtstamp.R
\name{extract_dtstamp}
\alias{extract_dtstamp}
\title{Extract the date-time stamp from a raw ELISA results file}
\usage{
extract_dtstamp(elisa_file_path)
}
\arguments{
\item{elisa_file_path}{full or relative path or name of ELISA results file}
}
\value{
POSIXct formatted date time stamp object
}
\description{
This function searches for and extracts the date-time stamp at which an
ELISA assay was read from a raw ELISA results file.
It uses a regular expression to find the character string containing the
date-time stamp; this string can include other text before the date.
The lubridate package is used to parse the dates - currently only European
and ISO date formats are included.
}
\note{
This function has not been tested extensively and may fail to identify some
date-time formats.

Pre-requisites:
\itemize{
\item ELISA reader raw results file for a 96-well plate format
\item Accepted file formats are .txt or .csv
\item The date is in European or ISO format (i.e. dmY or Ymd)
}
}
\examples{
 \dontrun{
 
 # Get date time stamp from an ELISA raw results file:
 elisa_dtstamp <- extract_dtstamp("Biotek_ex_20210922.csv")
 
 }

}
\author{
Amy Mikhail, \email{amy.mikhail@gmail.com}
}
