% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_ods.R
\name{extract_ods}
\alias{extract_ods}
\title{Extract optical densities from a raw ELISA results file}
\usage{
extract_ods(elisa_file_path)
}
\arguments{
\item{elisa_file_path}{full or relative path or name of ELISA results file}
}
\value{
data.table of ODs (8 rows A - H by 12 columns C01 - C12)
}
\description{
This function extracts optical densities (ODs) from any raw ELISA file.
It is platform independent and works by finding the block of data preceded
by the 8 row names (A - H) usually applied to a 96-well ELISA plate.
Currently the function does not support plates of a different size.
All header information and un-necessary columns are removed.

Pre-requisites:
\itemize{
\item ELISA reader raw results file for a 96-well plate format
\item Accepted file formats are .txt or .csv
\item The block of optical density results must be preceded by row labels A - H
}
}
\examples{
 \dontrun{
 
 # Import a .csv ELISA raw results file from a Biotek 800 TS reader:
 elisa_raw <- extract_ods("Biotek_ex_20210922.csv")
 
 }

}
\author{
Amy Mikhail, \email{amy.mikhail@gmail.com}
}
