% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/godata_getids.R
\name{godata_getids}
\alias{godata_getids}
\title{Go.Data get credentials for connection string}
\usage{
godata_getids(url = "http://localhost:3000/", ...)
}
\arguments{
\item{url}{Character string of Go.Data instance URL}

\item{username}{The email address the user uses to login to Go.Data}

\item{password}{The password the user uses to login to Go.Data}
}
\value{
list of one-time connection token, internal user and outbreak IDs
}
\description{
This function takes Go.Data user credentials as input and returns a one-time
connection token to the Go.Data database specified by the user credentials.
The function also returns the internal Go.Data ID for that user and the ID
of their active outbreak. These details are required every time the user
needs to import data from Go.Data to R or export data from R into Go.Data.
}
\note{
\enumerate{
\item The user must have the required permissions in their Go.Data account in
order to interact with the relevant Go.Data tables in R.  For example, if
the user wants to import or modify lab data, the function "lab manager" must
be included as one of their assigned roles in Go.Data.
\item It is strongly advised that user credentials be passed to this function
securely, e.g. instead of typing the password directly as a function
argument, use \code{\link[getPass:getPass]{getPass::getPass()}}.
\item The url for Go.Data instances that are off-line are usually composed of a
string like "http://localhost:3000/" (last component is the port number). On
-line Go.Data instances should have their own web address.
}

Pre-requisites:
\itemize{
\item The user must already have an account in the Go.Data database
\item The user must know their Go.Data user credentials
\item The user must know the url for their Go.Data instance
}
}
\examples{
 \dontrun{
 # Enter user credentials and check a token is returned:
 print(godata_getids(url = "http://localhost:3000/",
                          username = params$username, 
                          password = params$pwd))
 
 # Example use to import case data from Go.Data to R:
 cases <- httr::GET(paste0(url,
                           "api/outbreaks/",
                           godata_getids()$outbreak,
                           "/cases?"), 
                    add_headers(Authorization = paste("Bearer",
                                                      godata_getids()$token,
                                                      sep = " ")))
 }

}
\author{
Amy Mikhail, \email{amy.mikhail@gmail.com}
}
