% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/godata_getlablist.R
\name{godata_getlablist}
\alias{godata_getlablist}
\title{Fetch lab results from Go.Data in a flat table}
\usage{
godata_getlablist(godataids)
}
\arguments{
\item{godataids}{List object containing Go.Data url, outbreak ID and token}
}
\value{
data.table of flat lab results imported from Go.Data into R
}
\description{
This function imports lab results from Go.Data into R as a flat table (no
nesting, no sub-lists) to facilitate down-stream analysis.  Full credit to
James Fuller (WHO Go.Data) who wrote the import section of this function.
Variable sub-setting for line lists is specific to the MOL Go.Data instance.
}
\note{
Pre-requisites:
\itemize{
\item The user must already have an account in the Go.Data database
\item The user must know their Go.Data user credentials
\item The user must know the url for their Go.Data instance
}
}
\examples{
 \dontrun{
 # Fetch lab results:
 labres <- godata_getlablist(godataids = godata_getids())
 }

}
