% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/label_elisaresults.R
\name{label_elisaresults}
\alias{label_elisaresults}
\title{Label raw ELISA plate results with sample IDs}
\usage{
label_elisaresults(sampleids, elisaresults)
}
\arguments{
\item{sampleids}{Data.table of sample IDs in wide (plate) format}

\item{elisaresults}{Data.table of raw ELISA results in wide (plate) format}
}
\value{
Data.table of merged sample IDs and raw ELISA results
}
\description{
This function reshapes imported ELISA data and sample IDs from wide (plate
format) to long and merges the two data.tables together so that ELISA
results are associated with their sample IDs.
}
\note{
The function assumes column and row names are inherited from the import
function \code{\link[=extract_ods]{extract_ods()}} for 96-well ELISA plates.
Rows in the input tables should be labeled A:H
Columns in the input tables should be labeled C01:C12
}
\examples{
 \dontrun{
 # Label ELISA results with their sample IDs:
 elisadt <- label_elisaresults(sampleids = sampleids, 
                               elisaresults = elisa_raw)
 }

}
\author{
Amy Mikhail, \email{amy.mikhail@gmail.com}
}
