% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list2vector.R
\name{list2vector}
\alias{list2vector}
\title{Convert lists in a vector to their unlisted values}
\usage{
list2vector(dtcol)
}
\arguments{
\item{dtcol}{vector containing lists to unlist}
}
\value{
the vector with sub-lists converted to the values they contain
}
\description{
This function takes a vector containing lists with single values in each
slot as input, unlists them and returns the vector with the individual
values instead.
}
\note{
\enumerate{
\item Do not attempt to use this function with lapply as it will crash R!
\item Do not use on lists with more than one value; this will generate an error.
}
}
\examples{
 \dontrun{
 # Convert a vector containing sub-lists to the unlisted values:
 cases[, qanswers.sampledate := list2vector(qanswers.sampledate)]
 }
}
\author{
Amy Mikhail, \email{amy.mikhail@gmail.com}
}
