% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc_check.R
\name{qc_check}
\alias{qc_check}
\title{Check if quality control parameters have been met}
\usage{
qc_check(qctable)
}
\arguments{
\item{qctable}{Flextable object containing quality control summary.}
}
\value{
Quality control global result (pass or fail) in a character string
}
\description{
This function takes the results of the \code{\link[=runqc_...]{runqc_...()}} functions and checks
that all the main parameters have been met.  The binary result is used to
trigger other actions to complete the QC report (tick or cross mark if
passed or failed, respectively and advice on troubleshooting if failed.)
}
\examples{
 \dontrun{
 # Create final QC result:
 qc <- qc_check(qctable = qctab)
 }
}
\author{
Amy Mikhail, \email{amy.mikhail@gmail.com}
}
