% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_interactive.R
\name{run_interactive}
\alias{run_interactive}
\title{Launch interactive Shiny app}
\usage{
run_interactive()
}
\value{
QC report, lab results for validation, case line list
}
\description{
This function will launch an interactive Shiny app that
allows the user to set the parameters for the ELISA plate.
}
\examples{
\dontrun{
# Launch interactive dialogue box:
run_interactive()
}


}
\author{
Amy Mikhail \email{Amy.Mikhail@phe.gov.uk}
}
