% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runqc_ratios.R
\name{runqc_ratios}
\alias{runqc_ratios}
\title{Calculate quality control metrics for measles or rubella ELISA runs}
\usage{
runqc_ratios(elisadata, calinf, calref, ipcinf, ipcsup, incinf, incsup)
}
\arguments{
\item{elisadata}{Data.table of calculated ELISA ratios}

\item{calinf}{Lower threshold for kit calibrator}

\item{calref}{Reference value for the kit calibrator}

\item{ipcinf}{Lower threshold for kit positive control}

\item{ipcsup}{Upper threshold for kit positive control}

\item{incinf}{Lower threshold for kit negative control}

\item{incsup}{Upper threshold for kit negative control}
}
\value{
Flextable of quality control indicators for the ELISA run
}
\description{
This function determines whether kit and in-house controls for calculated
ratios from measles and rubella kits fall within the accepted range of
values defined by the manufacturer, thus validating an ELISA run.
If controls are duplicated, coefficients of variation are also calculated.
}
\note{
To be validated:
\itemize{
\item in-house and kit negative control ratios must be < 0.8
\item in-house and kit positive control ratios must be >= 1.1
\item kit control ratios must fall within limits set by EUROIMMUN for each lot
}
}
\examples{
 \dontrun{
 # Create QC table:
qctab <- runqc_ratios(elisadata = elisadt,
                      calinf = params$calinf,
                      calref = params$calref,
                      ipcinf = params$ipcinf,
                      ipcsup = params$ipcsup,
                      incinf = params$incinf,
                      incsup = params$incsup)
 }
}
\author{
Amy Mikhail, \email{amy.mikhail@gmail.com}
}
