% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runqc_standardcurve.R
\name{runqc_standardcurve}
\alias{runqc_standardcurve}
\title{Calculate quality control metrics for pertussis ELISA runs}
\usage{
runqc_standardcurve(
  elisadata,
  cal12inf,
  cal34inf,
  cipcinf,
  cipcsup,
  cincinf,
  cincsup
)
}
\arguments{
\item{elisadata}{Data.table of calculated ELISA values}

\item{cal12inf}{Lower limit (OD) for calibrators 1 - 2}

\item{cal34inf}{Lower limit (OD) for calibrators 3 - 4}

\item{cipcinf}{Lower threshold for kit positive control (IU/mL)}

\item{cipcsup}{Upper threshold for kit positive control (IU/mL)}

\item{cincinf}{Lower threshold for kit negative control (IU/mL)}

\item{cincsup}{Upper threshold for kit negative control (IU/mL)}
}
\value{
Flextable of quality control indicators for the ELISA run
}
\description{
This function determines whether kit controls and calibrators for standard
curve-derived values from pertussis ELISA kits fall within the accepted
range of values defined by the manufacturer, thus validating an ELISA run.
If controls are duplicated, coefficients of variation are also calculated.
}
\note{
To be validated:
\itemize{
\item kit negative controls must be < 40 IU/mL
\item kit positive controls must be >= 40 IU/mL
\item kit control values must fall within limits set by EUROIMMUN for each lot
}
}
\examples{
 \dontrun{
 # Create QC table:
qctab <- runqc_standardcurve(elisadata = elisadt,
                             cal12inf = params$cal12inf,
                             cal34inf = params$cal34inf,
                             cipcinf = params$cipcinf,
                             cipcsup = params$cipcsup,
                             cincinf = params$cincinf,
                             cincsup = params$cincsup)
 }
}
\author{
Amy Mikhail, \email{amy.mikhail@gmail.com}
}
